Bug 26684: Fix atomic update

+++ installer/data/mysql/atomicupdate/001-26684-drop_marc_column_from_auth_header.perl
@@ -1,6 +1,8 @@
 $DBversion = 'XXX'; # will be replaced by the RM
 if( CheckVersion( $DBversion ) ) {
-     $dbh->do( "ALTER TABLE auth_header DROP COLUMN marc" );
+    if( column_exists( 'auth_header', 'marc' ) ) {
+        $dbh->do( "ALTER TABLE auth_header DROP COLUMN marc" );
+    }
 
-    NewVersion( $DBversion, XXXXX, "Drop marc column from auth_header");
+    NewVersion( $DBversion, 26684, "Drop marc column from auth_header");
 }
